+++ crates/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # `fitstable-cli` Change Log
 
+## 0.1.1-beta
+
+Released 2026-03-04
+
+### Added
+
+* `hipsgen.bash` script to have a simple way to build a (small) HiPS^t catalog.
+
+### Fixed
+
+* In `mkhips`, the sorted FITS file and its HEALPix index do not have to be in the current directory anymore.
+
 ## 0.1.0-beta
 
 Released 2026-02-24